Seperating out `data_packet` from `data_emitter`
