+++ src/petalinux/apps/silayer-server/Makefile
@@ -4,6 +4,8 @@ TARGETS = vtctrl_server vtctrl_client silayer_server
 LD_FLAGS = -L$(HOME)/lib
 SERVER_LIBS = -lzmq -lvatactrl
 
+.PHONY: clean
+
 all: $(TARGETS)
 
 vtctrl_server: vtctrl_server.cpp
@@ -12,15 +14,21 @@ vtctrl_server: vtctrl_server.cpp
 vtctrl_client: vtctrl_client.cpp
 	g++ -o $@ $(CXX_FLAGS) $< -lzmq
 
-
 silayer_server.o: silayer_server.cpp
-	g++ -o $@ $(CXX_FLAGS) $<
+	g++ -o $@ $(CXX_FLAGS) -c $<
+
+silayer_server_main.o: silayer_server_main.cpp
+	g++ -o $@ $(CXX_FLAGS) -c $<
 
 data_emitter.o: data_emitter.cpp
-	g++ -o $@ $(CXX_FLAGS) $<
+	g++ -o $@ $(CXX_FLAGS) -c $<
 
-silayer_server: data_emitter.o silayer_server.o
-	g++ -o $@ $(CXX_FLAGS) $(LD_FLAGS) $< $(SERVER_LIBS)
+data_packet.o: data_packet.cpp
+	g++ -o $@ $(CXX_FLAGS) -c $<
+
+silayer_server: data_packet.o data_emitter.o silayer_server.o
+	g++ -o $@ $(LD_FLAGS) $< $(SERVER_LIBS)
 
 clean:
-	rm $(TARGETS) *~
+	$(RM) $(TARGETS) *.o *~
+
