+++ src/petalinux/apps/silayer-server/silayer_server_main.cpp
@@ -1,0 +1,8 @@
+#include "silayer_server.hpp"
+
+int main () {
+    LayerServer layer_server;
+    return layer_server.run();
+}
+
+// vim: set ts=4 sw=4 sts=4 et:
